feat(users): implement GET endpoints for user login by email across multiple services
