+++ src/app/api/EventOrganizer/users/[id]/route.ts
@@ -3,27 +3,6 @@ import { PrismaClient } from "@/../prisma/generated/postgresql2";
 
 const prisma = new PrismaClient();
 
-// ✅ Get User by ID (GET)
-export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
-  try {
-    const { id } = params;
-
-    const user = await prisma.eventOrganizerUser.findUnique({
-      where: { id },
-    });
-
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
-    }
-
-    return NextResponse.json(user, { status: 200 });
-
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
-  }
-}
-
 // ✅ Update User (PUT)
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
