+++ src/app/api/FoodieTraveler/users/[id]/route.ts
@@ -2,27 +2,6 @@ import { NextRequest, NextResponse } from "next/server";
 import { PrismaClient } from "@/../prisma/generated/postgresql";
 
 const prisma = new PrismaClient();
-// ✅ Get a User by ID (GET)
-export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
-  try {
-    const { id } = params;
-
-    const user = await prisma.userFoodie.findUnique({
-      where: { id },
-      select: { id: true, email: true, name: true, createdAt: true },
-    });
-
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
-    }
-
-    return NextResponse.json(user, { status: 200 });
-
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
-  }
-}
 
 // ✅ Update User Details (PUT)
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
