+++ src/app/api/GameHub/users/[id]/route.ts
@@ -3,24 +3,6 @@ import { PrismaClient as GameHubClient } from "@/../prisma/generated/postgresql2
 
 const prisma = new GameHubClient();
 
-export async function GET(_: NextRequest, { params }: { params: { id: string } }) {
-  try {
-    const user = await prisma.gameHubUser.findUnique({
-      where: { id: params.id },
-      select: { id: true, email: true, name: true, validated: true },
-    });
-
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
-    }
-
-    return NextResponse.json(user, { status: 200 });
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
-  }
-}
-
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
     const { name } = await req.json();
