+++ src/app/api/HealthTracker/users/[id]/route.ts
@@ -3,23 +3,6 @@ import { PrismaClient as HealthTrackerClient } from "@/../prisma/generated/postg
 
 const prisma = new HealthTrackerClient();
 
-export async function GET(_: NextRequest, { params }: { params: { id: string } }) {
-  try {
-    const user = await prisma.userHealthTracker.findUnique({
-      where: { id: params.id },
-    });
-
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
-    }
-
-    return NextResponse.json(user);
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
-  }
-}
-
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
     const { name } = await req.json();
