+++ src/app/api/PetLover/users/[id]/route.ts
@@ -1,50 +1,19 @@
 import { NextRequest, NextResponse } from "next/server";
 import { PrismaClient as PostgresqlClient } from "@/../prisma/generated/postgresql";
-import bcrypt from "bcryptjs";
 
 const prisma = new PostgresqlClient();
-const SALT_ROUNDS = 10;
-
-
-// ✅ Get User by ID (GET /{id})
-export async function GET(req: NextRequest) {
-  try {
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
-
-    if (!id) {
-      return NextResponse.json({ error: "User ID is required" }, { status: 400 });
-    }
-
-    const user = await prisma.userPetLover.findUnique({
-      where: { id },
-      select: { id: true, email: true, name: true, createdAt: true },
-    });
-
-    if (!user) {
-      return NextResponse.json({ error: "User not found" }, { status: 404 });
-    }
-
-    return NextResponse.json(user, { status: 200 });
-  } catch (error) {
-    console.error("Error fetching user:", error);
-    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
-  }
-}
 
 // ✅ Update User (PUT /{id})
-export async function PUT(req: NextRequest) {
+export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
     const { name } = await req.json();
 
-    if (!id || !name) {
+    if (!params.id || !name) {
       return NextResponse.json({ error: "User ID and name are required" }, { status: 400 });
     }
 
     const updatedUser = await prisma.userPetLover.update({
-      where: { id },
+      where: { id: params.id },
       data: { name },
     });
 
@@ -56,16 +25,14 @@ export async function PUT(req: NextRequest) {
 }
 
 // ✅ Delete User (DELETE /{id})
-export async function DELETE(req: NextRequest) {
+export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
 
-    if (!id) {
+    if (!params.id) {
       return NextResponse.json({ error: "User ID is required" }, { status: 400 });
     }
 
-    await prisma.userPetLover.delete({ where: { id } });
+    await prisma.userPetLover.delete({ where: { id: params.id } });
     return NextResponse.json({ message: "User deleted successfully" }, { status: 200 });
   } catch (error) {
     console.error("Error deleting user:", error);
