+++ src/app/api/PetLover/users/login/[email]/route.ts
@@ -1,0 +1,27 @@
+import { NextRequest, NextResponse } from "next/server";
+import { PrismaClient as PostgresqlClient } from "@/../prisma/generated/postgresql";
+
+const prisma = new PostgresqlClient();
+
+// ✅ Get User by ID (GET /{id})
+export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
+  try {
+    if (!params.id) {
+      return NextResponse.json({ error: "User ID is required" }, { status: 400 });
+    }
+
+    const user = await prisma.userPetLover.findUnique({
+      where: { id: params.id },
+      select: { id: true, email: true, name: true, createdAt: true },
+    });
+
+    if (!user) {
+      return NextResponse.json({ error: "User not found" }, { status: 404 });
+    }
+
+    return NextResponse.json(user, { status: 200 });
+  } catch (error) {
+    console.error("Error fetching user:", error);
+    return NextResponse.json({ error: "Failed to fetch user" }, { status: 500 });
+  }
+}
