+++ src/app/api/SocialButterfly/users/[id]/route.ts
@@ -3,30 +3,6 @@ import { PrismaClient as PostgresqlClient } from '@/../prisma/generated/postgres
 
 const prisma = new PostgresqlClient()
 
-// GET user by ID
-export async function GET(_: NextRequest, { params }: { params: { id: string } }) {
-  try {
-    const user = await prisma.userSocialButterfly.findUnique({
-      where: { id: params.id },
-      select: {
-        id: true,
-        email: true,
-        name: true,
-        createdAt: true
-      }
-    })
-
-    if (!user) {
-      return NextResponse.json({ error: 'User not found' }, { status: 404 })
-    }
-
-    return NextResponse.json(user, { status: 200 })
-  } catch (err) {
-    console.error(err)
-    return NextResponse.json({ error: 'Server error' }, { status: 500 })
-  }
-}
-
 // PUT update user name
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
