+++ src/app/api/VolunteerOrg/users/[id]/route.ts
@@ -3,27 +3,6 @@ import { PrismaClient as PostgresqlClient } from '@/../prisma/generated/postgres
 
 const prisma = new PostgresqlClient()
 
-export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
-  try {
-
-    const { id } = params
-    const user = await prisma.userVolunteerOrg.findUnique({ where: { id } })
-
-    if (!user) {
-      return NextResponse.json({ error: 'User not found' }, { status: 404 })
-    }
-
-    return NextResponse.json(user)
-
-  } catch (err) {
-    let errMessage = 'Internal server error'
-    if (err instanceof Error) {
-      errMessage = err.message
-    }
-    return NextResponse.json({ error: errMessage }, { status: 500 })
-  }
-}
-
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
     const { id } = params
